Move Sanitizer Outside of ExportersDirectory
